[level 0] Title: 개미 군단, Time: 0.02 ms, Memory: 79.9 MB -BaekjoonHub

+++ 프로그래머스/0/120837. 개미 군단/개미 군단.java
@@ -1,5 +1,5 @@
 class Solution {
     public int solution(int hp) {
-        return (int)hp/5 + (int)hp%5/3 + (hp%5)%3 ;
+        return hp/5 + hp%5/3 + hp%5%3 ;
     }
 }
